인증, CORS 추가

+++ routes/events.py
@@ -1,11 +1,12 @@
-from fastapi import APIRouter, status, HTTPException
+from typing import List
+from beanie import PydanticObjectId
+from fastapi import APIRouter, status, HTTPException, Depends
+
 from models.events import Event, EventUpdate
 from database.connection import Database
-from beanie import PydanticObjectId
-from typing import List
+from auth.authenticate import authenticate
 
 event_router = APIRouter(tags=["Events"])
-
 event_database = Database(Event)
 
 
@@ -27,13 +28,20 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user:str=Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {"message": "Event created sucessfully."}
 
 
 @event_router.delete("/delete/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
+async def delete_event(id: PydanticObjectId, user:str=Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Operation not allowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
@@ -45,8 +53,14 @@ async def delete_event(id: PydanticObjectId) -> dict:
 
 @event_router.put("/edit/{id}", response_model=Event)
 async def update_event(
-    id: PydanticObjectId, body:EventUpdate 
+    id: PydanticObjectId, body:EventUpdate, user:str=Depends(authenticate) 
 ) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not update_event:    
         raise HTTPException(
